automation/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-genai
python-dotenv
beautifulsoup4
dateparser
pytz
